requirements.txt
requests==2.25.1
pydantic==1.8.1
fastapi==0.63.0
python-dotenv==0.15.0
